backend/requirements.txt
django==4.0
djoser==2.1.0
djangorestframework==3.12.4
djangorestframework_simplejwt==4.8.0
pytz==2021.3
